+++ ignite/templates/field/field.go
@@ -27,6 +27,17 @@ func (f Field) DataType() string {
 	return dt.DataType(f.Datatype)
 }
 
+// IsSlice returns true if the field is a slice.
+// It assumes that a non indexable type is a slice.
+func (f Field) IsSlice() bool {
+	dt, ok := datatype.IsSupportedType(f.DatatypeName)
+	if !ok {
+		panic(fmt.Sprintf("unknown type %s", f.DatatypeName))
+	}
+
+	return dt.NonIndex
+}
+
 // ProtoFieldName returns the field name used in proto.
 func (f Field) ProtoFieldName() string {
 	return f.Name.Snake